# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Confecciones WA",
        "description": "Documentación de la API para el sistema de Confecciones WA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "description": "Obtiene el listado paginado de usuarios con sus datos completos y roles.",
                "operationId": "dbe72d85943ad52189a017183632a4dd",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombres, apellidos, documento o correo.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Juan"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10,
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuarios obtenidos correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registrar usuario",
                "description": "Crea un nuevo usuario con todos sus datos y un rol.",
                "operationId": "82b7b563affba6ec7aefa2bba435f41e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "is_active"
                                ],
                                "properties": {
                                    "document_type_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "document_number": {
                                        "type": "string",
                                        "example": "12345678",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Carlos",
                                        "nullable": true
                                    },
                                    "last_name_father": {
                                        "type": "string",
                                        "example": "Perez",
                                        "nullable": true
                                    },
                                    "last_name_mother": {
                                        "type": "string",
                                        "example": "Gomez",
                                        "nullable": true
                                    },
                                    "gender_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@correo.com",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "rol_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "description": "true activo, false inactivo",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuarios/{id}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario",
                "description": "Actualiza la información completa de un usuario y su rol.",
                "operationId": "0bde4772cb746960ac7cfdedda7ee90c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_active"
                                ],
                                "properties": {
                                    "document_type_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "document_number": {
                                        "type": "string",
                                        "example": "12345678",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Carlos",
                                        "nullable": true
                                    },
                                    "last_name_father": {
                                        "type": "string",
                                        "example": "Perez",
                                        "nullable": true
                                    },
                                    "last_name_mother": {
                                        "type": "string",
                                        "example": "Gomez",
                                        "nullable": true
                                    },
                                    "gender_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@correo.com",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456",
                                        "nullable": true
                                    },
                                    "rol_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "description": "true activo, false inactivo",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}